AUTOACH-18 Status Checkers should return a list of response records for each external ACH ID instead of a single response record.
It is possible for a provider to return multiple responses for the same ACH, even
for polling that happens with a date range of no more than 24 hours.
Previously, we merged the response hashes such that responses with the same
key (external ach id) would be overwritten. Now, the response hashes have
an Array of ACH responses instead of a single ACH response. When merging these
hashes, we concatenate the values instead of overwriting them.

Unfortunately this is a breaking change, so major version has been updated (1.0.0)

Also upgraded to Ruby 2.5.0 (for gem development)